devide message.hpp
