+++ hermesnet/hermes/message/message_generator.cpp
@@ -1,0 +1,10 @@
+
+#include "message_generator.h"
+
+using namespace network::message::v2;
+
+template <typename T>
+void MessageGenerator::generate()
+{
+    T type;
+}
